requirements_dev.txt
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0
Click==7.0
pytest==4.6.5
pytest-runner==5.1
deltalake>=0.2.1
pyspark==3.0.1
pytest-spark==0.6.0
neo4j
pandas>=1.1.0
PyYAML>=5.4
yamale>=3.0.4
jsonpath-ng>=1.5.2